+++ src/database/init.rs
@@ -1,7 +1,7 @@
 //! Database initialization
 
-use std::path::PathBuf;
 use std::io;
+use std::path::PathBuf;
 use tokio::fs;
 
 use crate::database::Database;
@@ -15,22 +15,20 @@ impl Database {
         match fs::remove_dir_all(&path).await {
             Ok(_) => {
                 log::warn!("the old database has been deleted, creating...");
-            },
+            }
             Err(e) => {
                 if e.kind() == io::ErrorKind::NotFound {
                     log::warn!("the database does not exist, creating...");
                 } else {
                     return Err(e);
                 }
-            },
+            }
         };
 
         // And create an empty new one
         fs::create_dir(&path).await?;
         log::info!("the database has been created");
 
-        Ok(Self {
-            path,
-        })
+        Ok(Self { path })
     }
 }