+++ src/database/table.rs
@@ -6,14 +6,13 @@ use crate::database::Database;
 
 #[derive(Debug)]
 pub struct Table {
-    pub name: String, 
+    pub name: String,
     pub file: PathBuf,
 }
 
 impl Database {
     /// Creates a new table in the database
     pub fn table(&self, name: &str) -> Table {
-
         let mut file = PathBuf::from(self.path.as_path());
         file.push(&name);