+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.7 - 2022-01-28
+
+### Changed
+
+- Remove `thiserror` crate dependency
+- `telegram` module
+
+## 0.1.6 - 2022-01-27
+
+### Changed
+
+- `digitalocean` module
+- `database` module
+- `telegram` module
+- `worker` module
+
 ## 0.1.5 - 2022-01-26
 
 ### Added